+++ Changelog.md
@@ -10,6 +10,7 @@
   * HTTP transport bot-nick parameter added to config.
   * ybot_manager get_all_transports_pid api added.
   * Use cowboy web server instead inets httpd for http serving.
+  * Added http `PUT` request support. Resend request body from http to all runned transports.
 
 ## Ybot-0.3 --> released